requirements.txt
requests
selenium
beautifulsoup4
pandas
python-dotenv
